+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.2.0] - 2024-05-01
+
+### Added
+
+- a `--mysql` option to use mysql instead sqlite default one
+
+## [0.1.4] - 2024-04-30
+
+### Changed
+
+- Add some informations on README
 
 ## [0.1.3] - 2024-04-30